The roll distribution

Density, distribution function, quantile function, and random generation for the discrete distribution described by a roll expression. See below for more details.

roll(roll, precise = FALSE)

droll(x, roll)

proll(q, roll, lower.tail = TRUE)

qroll(p, roll, lower.tail = TRUE)

rroll(n, roll, verbose = TRUE)

Arguments

roll
A roll expression (e.g., 2 * d6 + 5).

precise
Whether to return values with arbitrary precision.

x
A numeric vector of outcomes.

q
A numeric vector of outcomes.

lower.tail
Whether to calculate P[X <= x] or P[X > x].

p
A numeric vector of probabilities.

n
Number of random deviates to return.

verbose
For rroll(), if n = 1, whether to print partial result of the dice roll.

Source

The main algorithm for calculating dice probabilities comes from MathWorld.

Symbolic calculations are handled by Ryacas, and, by extension, by Yacas.

Value

For roll(), a data frame with three columns: outcome, count, and freq. For droll(), proll(), qroll(), and rroll(), a numeric vector.

Details

Given a roll expression (i.e., an arithmetic expression involving dice), roll() calculates the complete distribution of the outcomes. This is possible because the distribution is discrete and has a finite number of outcomes.

From this distribution, droll() returns the density, proll() returns the distribution function, qroll() returns the quantile function, and rroll() generates random deviates. They mirror functions from the Distributions family.

Roll Expressions

A roll expression is a piece of R code that describes a dice roll with or without modifiers, e.g., 2 * d6 + 5.

Standard dice notation should mostly work out of the box, with the notable exception of NdX, i.e., "roll N dice with X faces and add the results". In this case, the user must write N * dX; this also means that, when translating "roll a die with X faces and multiply the result by N" to a roll expression, the user must then write dX * N. All other expressions involving dice can usually be pasted straight into these functions.

For more information, see the dice creating function d().

Built-in Dice

It is possible to define any die with d(), but some are already built-in. Because of R's restrictions on what kind of object can be exported, they are not readly available for the user, but can be used inside a roll expression nontheless. These are the standard D&D dice: d4, d6, d8, d10, d12, d20, and d100.

Arbitrary Precision

Most dice programs that can calculate probabilities are forced to round their results due to the fact that these quantities might become exceptionally low when dealing with a lot of dice. This, however, can lead to error magnification.

In order to avoid rouding as much as possible, all functions described here use Ryacas::yac_str() to run computations symbolically. By default, results are converted to numeric vectors just before returning to the user, but one is able to access the symbolic strings returned by Ryacas by setting precise = TRUE on roll().

Examples

# Complete distribution of 2d6 + 5
roll(2 * d6 + 5)

#> # A tibble: 11 × 3
#>    outcome count   freq
#>      <dbl> <dbl>  <dbl>
#>  1       7     1 0.0278
#>  2       8     2 0.0556
#>  3       9     3 0.0833
#>  4      10     4 0.111 
#>  5      11     5 0.139 
#>  6      12     6 0.167 
#>  7      13     5 0.139 
#>  8      14     4 0.111 
#>  9      15     3 0.0833
#> 10      16     2 0.0556
#> 11      17     1 0.0278

# Density of 2d6 + 5
droll(12, 2 * d6 + 5)

#> [1] 0.1666667

# Distribution function of 2d6 + 5
proll(12, 2 * d6 + 5)

#> [1] 0.5833333

# Quantile function of 2d6 + 5
qroll(0.5, 2 * d6 + 5)

#> [1] 12

# Roll 2d6 + 5 (generate random deviates)
set.seed(42)
rroll(1, 2 * d6 + 5)

#> # Outcome: sum(1, 5) + 5 
#> 

#> [1] 11